+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the "vscode-random" extension are documented in this file
 ## Added
 - New command to allow random seed reset
 
+## Changed
+- Refactoring to improve code simplicity and readability
+
 ## [1.0.2] - 2017-01-04
 ### Fixed
 - Fix bug when generating random values with multiple editors